new handling for pysolc to ensure better compatibility
