+++ Dockerfile
@@ -31,8 +31,8 @@ RUN apt-get update && apt-get -y install\
 COPY requirements.txt /tmp/
 RUN pip3 install --user -r /tmp/requirements.txt
 
-COPY scripts/isolc/ /isolc/
-RUN cd / && python3 -m isolc.install_solc
+COPY scripts /tmp/installsolc
+RUN cd /tmp/ && python3 -m installsolc.install_solc
 
 # install truffle for project compilation
 RUN apt-get update && apt-get install -y\
