+++ scripts/install_solc.py
