+++ scripts/solc_project.py
@@ -27,16 +27,8 @@ from solcx.exceptions import SolcError
 from solcx.main import _parse_compiler_output
 from solcx.wrapper import solc_wrapper
 
-from . import project, utils
-
-
-def _get_binary(solc_version):
-    """Returns the binary for some version of solc.
-    """
-    binary = os.path.join(get_solc_folder(), f'solc-v{solc_version}')
-    assert os.path.exists(binary), 'solc binary not found'
-    return binary
-
+from . import project
+from . import pysolc
 
 class SolcProject(project.Project):
     """A project that uses the `solc` compiler.
@@ -48,7 +40,7 @@ class SolcProject(project.Project):
         sources = self._get_sol_files()
 
         if not sources:
-            raise utils.NoSolidityProject(self.project_root)
+            raise pysolc.NoSolidityProject(self.project_root)
 
         comp_output = self._compile_solfiles(sources)
 
@@ -64,38 +56,9 @@ class SolcProject(project.Project):
                 '/test/' not in p[len(str(self.project_root)):] and
                 not p.endswith('/test')]
 
-    def _compile_solfiles(self, files, solc_version=None, output_values=utils.OUTPUT_VALUES):
+    def _compile_solfiles(self, files, solc_version=None, output_values=pysolc.OUTPUT_VALUES):
         """Compiles the files using the solc compiler.
         """
-        node_modules_dir = utils.find_node_modules_dir(self.project_root)
+        pysolc.compile_solfiles(files, self.project_root, solc_version, output_values)
 
-        remappings = []
 
-        if node_modules_dir is not None:
-            zeppelin_path = os.path.abspath(os.path.join(
-                node_modules_dir, 'zeppelin-solidity'))
-            open_zeppelin_path = os.path.abspath(
-                os.path.join(node_modules_dir, 'openzeppelin-solidity'))
-            if os.path.isdir(zeppelin_path):
-                remappings.append(f'zeppelin-solidity={zeppelin_path}')
-            if os.path.isdir(open_zeppelin_path):
-                remappings.append(f'openzeppelin-solidity={open_zeppelin_path}')
-
-        if solc_version is None:
-            solc_version = min(map(utils.parse_sol_version, files),
-                               key=utils.version_to_tuple)
-
-        binary = _get_binary(solc_version)
-
-        combined_json = ','.join(output_values)
-        compiler_kwargs = {'import_remappings': remappings,
-                           'allow_paths': self.project_root,
-                           'source_files': files,
-                           'solc_binary': binary,
-                           'combined_json': combined_json}
-
-        try:
-            stdoutdata, _, _, _ = solc_wrapper(**compiler_kwargs)
-            return _parse_compiler_output(stdoutdata)
-        except SolcError as e:
-            raise utils.SolidityCompilationException(e, files)
